+++ src/components/common/Modal.tsx
@@ -1,9 +1,11 @@
+'use client'
+
 import { type ReactNode } from 'react'
 import FocusLock from 'react-focus-lock'
 
 import CloseIcon from '@/assets/icons/modal/close.svg'
 import Portal from '@/components/common/modal/Portal'
-import PORTAL_CONTAINER_ID from '@/constants/portalContainerId'
+import { PORTAL_CONTAINER_ID } from '@/constants'
 import { Z_INDEX } from '@/foundations/zIndex'
 import { useModal } from '@/hooks/useModal'
 import { cn } from '@/utils/cn'
