+++ src/components/common/index.ts
@@ -1,0 +1,12 @@
+export { default as Button } from './Button'
+export { default as Card } from './Card'
+export { default as Input } from './Input'
+export { default as Modal } from './Modal'
+export { default as Carousel } from './Carousel'
+export { default as Dropdown } from './Dropdown'
+export { default as GaugeBar } from './GaugeBar'
+export { default as ItemRowContent } from './ItemRowContent'
+export { default as Pagination } from './Pagination'
+export { default as Slider } from './Slider'
+export { default as Tag } from './Tag'
+export { default as StarRating } from './StarRating'
