+++ src/components/layout/index.ts
@@ -1,0 +1,4 @@
+export { default as Header } from './Header'
+export { default as Footer } from './Footer'
+export { default as FooterLogo } from './FooterLogo'
+export { default as RenderInfo } from './RenderInfo'
